Basic instruction decode test

+++ src/cpu.rs
@@ -39,7 +39,7 @@ impl Cpu {
 
         if pc == self.registers.PC {
             // If we did not execute a jump, proceed to next instruction as usual
-            self.registers.PC = pc + size;
+            self.registers.PC = pc + size as u16;
             cycles.not_taken()
         } else {
             cycles.taken()
